requirements.txt
# DB Migrator Dependencies
streamlit>=1.30.0
streamlit-javascript>=0.1.5
psycopg2-binary>=2.9.9
pandas>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
